server: renamed the search module
