+++ server/services/searchService.go
@@ -1,4 +1,4 @@
-package search
+package services
 
 import (
 	"encoding/json"
